Move exceptions to core module
