+++ grid/app/main/sfl/processes/process_manager.py
@@ -5,12 +5,7 @@ from .config import Config
 # PyGrid imports
 from ...storage.warehouse import Warehouse
 from ..syft_assets import plans, protocols
-from ...exceptions import (
-    FLProcessConflict,
-    ProcessNotFoundError,
-    PlanNotFoundError,
-    ProtocolNotFoundError,
-)
+from ...core.exceptions import FLProcessConflict, ProcessNotFoundError, PlanNotFoundError, ProtocolNotFoundError
 
 
 class ProcessManager:
